+++ backend/app/models/word.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BIGINT, Enum, ForeignKey, String
+from sqlalchemy import Integer, Enum, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.models.base import Base, TimestampMixin
@@ -8,8 +8,8 @@ from app.models.enums import TagType, WordType
 class Word(TimestampMixin, Base):
     __tablename__ = "word"
 
-    id: Mapped[int] = mapped_column(BIGINT, primary_key=True, autoincrement=True)
-    unit_id: Mapped[int] = mapped_column(BIGINT, ForeignKey("unit.id", ondelete="CASCADE"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    unit_id: Mapped[int] = mapped_column(Integer, ForeignKey("unit.id", ondelete="CASCADE"), nullable=False)
     english: Mapped[str] = mapped_column(String(500), nullable=False)
     chinese: Mapped[str] = mapped_column(String(500), nullable=False)
     type: Mapped[WordType] = mapped_column(Enum(WordType), nullable=False, default=WordType.word)
@@ -26,7 +26,7 @@ class WordTag(Base):
     __tablename__ = "word_tags"
 
     word_id: Mapped[int] = mapped_column(
-        BIGINT, ForeignKey("word.id", ondelete="CASCADE"), primary_key=True
+        Integer, ForeignKey("word.id", ondelete="CASCADE"), primary_key=True
     )
     tag: Mapped[TagType] = mapped_column(Enum(TagType), primary_key=True)
